requirements.txt
h5py==2.9.0
ortools==8.0.8283
pyyaml==5.1.2
numpy==1.17.3
torch==1.1.0
torch-scatter==1.4.0
torch-sparse==0.4.3
torch-cluster==1.4.5
torch-geometric==1.3.2

https://pytorch-geometric.com/whl/torch-1.7.0.html

wget https://pytorch-geometric.com/whl/torch-1.7.0/torch_scatter-2.0.5%2Bcu102-cp37-cp37m-linux_x86_64.whl
pip3 install torch_scatter-2.0.5+cu102-cp37-cp37m-linux_x86_64.whl
wget https://pytorch-geometric.com/whl/torch-1.7.0/torch_sparse-0.6.8%2Bcu102-cp37-cp37m-linux_x86_64.whl
pip3 install torch_sparse-0.6.8+cu102-cp37-cp37m-linux_x86_64.whl

wget https://pytorch-geometric.com/whl/torch-1.7.0/torch_cluster-1.5.8%2Bcu102-cp37-cp37m-linux_x86_64.whl
pip3 install torch_cluster-1.5.8+cu102-cp37-cp37m-linux_x86_64.whl
wget https://pytorch-geometric.com/whl/torch-1.7.0/torch_spline_conv-1.2.0%2Bcu102-cp37-cp37m-linux_x86_64.whl
pip3 install torch_spline_conv-1.2.0+cu102-cp37-cp37m-linux_x86_64.whl
pip3 install torch-geometric
